requirements.dev.txt
ipython>=7.0.0,<=8.0.0
ipdb
pudb
pytest-watch

# install project as editable
-e .
